move tests and typings to the top. Fix unit tests, add tslint
