+++ e2e/assets/yellow.gif
