+++ e2e/webpack.config.js
@@ -1,5 +1,6 @@
 module.exports = {
   entry: './inject.js',
+  mode: 'development',
   output: {
     filename: './build/inject.bundle.js',
     library: 'body2sketch'
