--- html2asketch/options.d.ts
@@ -1,8 +1,0 @@
-export default interface Options {
-  addArtboard?: boolean,
-  artboardName?: string,
-  pageName?: string,
-  getRectangleName?: (node: HTMLElement) => string,
-  skipSystemFonts?: boolean
-}
-
