+++ tests/helpers/text.spec.js
@@ -1,4 +1,4 @@
-import {fixWhiteSpace} from './text';
+import {fixWhiteSpace} from '../../html2asketch/helpers/text';
 
 // Sequences of whitespace are collapsed. Newline characters in the source are handled the same as other whitespace.
 test('normal & nowrap', () => {
