+++ tslint.json
@@ -1,0 +1,12 @@
+{
+    "defaultSeverity": "error",
+    "extends": [
+        "tslint:recommended"
+    ],
+    "jsRules": {},
+    "rules": {
+      "interface-name": [true, "never-prefix"],
+      "variable-name": [true, "allow-leading-underscore"]
+    },
+    "rulesDirectory": []
+}
