+++ typings/model/artboard.d.ts
@@ -1,4 +1,4 @@
-import Base from './base';
+import Base from "./base";
 
 export interface ArtboardProps {
   width: number;
