+++ typings/model/base.d.ts
