+++ typings/model/document.d.ts
