+++ typings/model/style.d.ts
