+++ typings/model/svg.d.ts
@@ -1,4 +1,4 @@
-import Base from './base';
+import Base from "./base";
 
 export interface SVGProps {
   x: number;
