+++ typings/model/symbolInstance.d.ts
@@ -1,4 +1,4 @@
-import Base from './base';
+import Base from "./base";
 
 export interface SymbolInstanceProps {
   x: number;
