+++ typings/model/textStyle.d.ts
