+++ typings/nodeTreeToSketchGroup.d.ts
@@ -1,4 +1,4 @@
-import Group from './model/group';
+import Group from "./model/group";
 import Options from "./options";
 
 export default function nodeTreeToSketchGroup(node: HTMLElement, options?: Options): Group;
