+++ typings/nodeTreeToSketchPage.d.ts
@@ -1,4 +1,4 @@
-import Page from './model/page';
+import Page from "./model/page";
 import Options from "./options";
 
 export default function nodeTreeToSketchPage(node: HTMLElement, options?: Options): Page;
