+++ typings/options.d.ts
@@ -1,0 +1,7 @@
+export default interface Options {
+  addArtboard?: boolean;
+  artboardName?: string;
+  pageName?: string;
+  getRectangleName?: (node: HTMLElement) => string;
+  skipSystemFonts?: boolean;
+}
